docker: avoid problem with tls site
This is a temporarily solution. Or should be.

TLS site is not opening anymore because for some reason it asks for
version 6.0.11 while server - same one that was used to deploy it - has
version 6.0.111 installed.

I don't have time to solve this right now, it is more than 1am and I
have to wake up 7am to work.